+++ AlchemistCode/Cards/Common/Fizz.cs
@@ -1,5 +1,4 @@
 using Alchemist.AlchemistCode.Commands;
-using BaseLib.Utils;
 using MegaCrit.Sts2.Core.Entities.Cards;
 using MegaCrit.Sts2.Core.GameActions.Multiplayer;
 
@@ -12,13 +11,13 @@ public class Fizz : AlchemistCard
 
     public Fizz() : base(1, CardType.Skill, CardRarity.Common, TargetType.Self)
     {
-        WithCards(1, 0);
+        WithUpgradingCardTip<Token.BurstingMix>();
         WithUpgradingCardTip<Token.ZestyMix>();
     }
 
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {
-        await CommonActions.Draw(this, choiceContext);
+        await Mixing.CreateOne<Token.BurstingMix>(choiceContext, Owner, IsUpgraded, this);
         await Mixing.CreateOne<Token.ZestyMix>(choiceContext, Owner, IsUpgraded, this);
     }
 }